Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyPDF2
chromadb
sentence-transformers
requests
